python/requirements.txt
langchain-community
langchain-openai
langchain_iris
fastapi[standard]
streamlit
mistral_common
langchain_mistralai
../jupyter/intersystems_irispython-3.2.0-py3-none-any.whl